+++ D1Data/D1EventManager.h
@@ -31,14 +31,14 @@ namespace BnD {
         virtual int64 implMakeNewSerialNumber();
     protected:
         int64 makeNewSerialNumber();
-        bool getTaskTransferInfo(int32 taskID, B1String* commandID, B1String* carrierID);
+        bool getTaskTransferInfo(int64 taskID, B1String* commandID, B1String* carrierID);
     public:
         void initialize(D1RedisClientInterface* redisClientInterface);
         void finalize();
-        bool addEvent(int32 code, const B1String& commandID, const B1String& carrierID, int32 taskID, int32 zoneID, int32 reason, bool isEssential = false);
+        bool addEvent(int32 code, const B1String& commandID, const B1String& carrierID, int64 taskID, int32 zoneID, int32 reason, bool isEssential = false);
         bool addEvent(int32 code);
         bool addEvent(int32 code, int32 zoneID);
-        bool addEvent(int32 code, int32 taskID, int32 zoneID, int32 reason, bool isEssential = false);
+        bool addEvent(int32 code, int64 taskID, int32 zoneID, int32 reason, bool isEssential = false);
     };
 }   //  !BnD
